// src/services/api/models/sale-with-assets.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Sale } from './sale';
import { SaleAsset } from './sale-asset';
/**
 * 
 * @export
 * @interface SaleWithAssets
 */
export interface SaleWithAssets {
    /**
     * 
     * @type {Sale}
     * @memberof SaleWithAssets
     */
    sale?: Sale;
    /**
     * 
     * @type {Array<SaleAsset>}
     * @memberof SaleWithAssets
     */
    saleAssets?: Array<SaleAsset> | null;
}