// src/services/api/models/sale.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SaleAssetType } from './sale-asset-type';
import { SaleStatus } from './sale-status';
import { SaleType } from './sale-type';
/**
 * 
 * @export
 * @interface Sale
 */
export interface Sale {
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Sale
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof Sale
     */
    created?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Sale
     */
    updated?: Date;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    livestreamId?: string | null;
    /**
     * 
     * @type {SaleType}
     * @memberof Sale
     */
    saleType?: SaleType;
    /**
     * 
     * @type {SaleStatus}
     * @memberof Sale
     */
    saleStatus?: SaleStatus;
    /**
     * 
     * @type {SaleAssetType}
     * @memberof Sale
     */
    saleAssetType?: SaleAssetType;
    /**
     * 
     * @type {number}
     * @memberof Sale
     */
    floorPrice?: number;
    /**
     * 
     * @type {number}
     * @memberof Sale
     */
    quantity?: number;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    saleParentId?: string;
    /**
     * 
     * @type {Date}
     * @memberof Sale
     */
    startTime?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof Sale
     */
    endTime?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    winningBuyOrderId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    sellerUserId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    categoryId?: string;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    communityId?: string;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    assetPackId?: string;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    assetId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Sale
     */
    isListedOnMarketplace?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Sale
     */
    isSellerPaid?: boolean;
}