// src/services/api/models/wallet-asset.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Blockchain } from './blockchain';
import { RedemptionStatus } from './redemption-status';
import { TransferStatus } from './transfer-status';
import { TransferWallet } from './transfer-wallet';
/**
 * 
 * @export
 * @interface WalletAsset
 */
export interface WalletAsset {
    /**
     * 
     * @type {string}
     * @memberof WalletAsset
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof WalletAsset
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof WalletAsset
     */
    created?: Date;
    /**
     * 
     * @type {Date}
     * @memberof WalletAsset
     */
    updated?: Date;
    /**
     * 
     * @type {Blockchain}
     * @memberof WalletAsset
     */
    blockchain?: Blockchain;
    /**
     * 
     * @type {string}
     * @memberof WalletAsset
     */
    ownerAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WalletAsset
     */
    contractAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WalletAsset
     */
    tokenId?: string | null;
    /**
     * 
     * @type {RedemptionStatus}
     * @memberof WalletAsset
     */
    redemptionStatus?: RedemptionStatus;
    /**
     * 
     * @type {boolean}
     * @memberof WalletAsset
     */
    isTransferable?: boolean;
    /**
     * 
     * @type {string}
     * @memberof WalletAsset
     */
    userId?: string;
    /**
     * 
     * @type {TransferStatus}
     * @memberof WalletAsset
     */
    transferStatus?: TransferStatus;
    /**
     * 
     * @type {TransferWallet}
     * @memberof WalletAsset
     */
    transferWallet?: TransferWallet;
    /**
     * 
     * @type {string}
     * @memberof WalletAsset
     */
    transferWalletAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WalletAsset
     */
    saleId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WalletAsset
     */
    assetPackId?: string;
}